requirements.txt
numpy
Pillow
pyquaternion
html-testRunner
setuptools
natsort
tqdm